+++ packages/boilerplate/theme/nuxt.config.js
@@ -93,5 +93,14 @@ export default {
         })
       })
     ]
+  },
+  router: {
+    scrollBehavior (_to, _from, savedPosition) {
+      if (savedPosition) {
+        return savedPosition;
+      } else {
+        return { x: 0, y: 0 };
+      }
+    }
   }
 };