+++ packages/commercetools/api-client/src/api/customerChangeMyPassword/index.ts
@@ -1,9 +1,10 @@
-import { apolloClient } from '../../index';
+import { getSettings } from '../../index';
 import CustomerChangeMyPassword from './defaultMutation';
 import { ChangeMyPasswordResponse } from '../../types/Api';
 
 const customerChangeMyPassword = async (version: any, currentPassword: string, newPassword: string): Promise<ChangeMyPasswordResponse> => {
-  return await apolloClient.mutate({
+  const { client } = getSettings();
+  return await client.mutate({
     mutation: CustomerChangeMyPassword,
     variables: {
       version,