+++ packages/commercetools/theme/components/MyAccount/BillingAddressForm.vue
@@ -1,12 +1,12 @@
 <template>
   <ValidationObserver v-slot="{ handleSubmit }">
     <form
-      id="shipping-details-form"
+      id="billing-details-form"
       class="form"
       @submit.prevent="handleSubmit(submitForm)"
     >
       <SfCheckbox
-        data-cy="shipping-details-checkbox_isDefault"
+        data-cy="billing-details-checkbox_isDefault"
         v-model="form.isDefault"
         name="isDefault"
         label="Set as default"
@@ -19,7 +19,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_firstName"
+            data-cy="billing-details-input_firstName"
             v-model="form.firstName"
             name="firstName"
             label="First Name"
@@ -34,7 +34,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_lastName"
+            data-cy="billing-details-input_lastName"
             v-model="form.lastName"
             name="lastName"
             label="Last Name"
@@ -50,7 +50,7 @@
         class="form__element"
       >
         <SfInput
-          data-cy="shipping-details-input_streetName"
+          data-cy="billing-details-input_streetName"
           v-model="form.streetName"
           name="streetName"
           label="Street Name"
@@ -60,7 +60,7 @@
         />
       </ValidationProvider>
       <SfInput
-        data-cy="shipping-details-input_apartment"
+        data-cy="billing-details-input_apartment"
         v-model="form.apartment"
         name="apartment"
         label="House/Apartment number"
@@ -74,7 +74,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_city"
+            data-cy="billing-details-input_city"
             v-model="form.city"
             name="city"
             label="City"
@@ -89,7 +89,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_state"
+            data-cy="billing-details-input_state"
             v-model="form.state"
             name="state"
             label="State/Province"
@@ -106,7 +106,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_zipCode"
+            data-cy="billing-details-input_zipCode"
             v-model="form.postalCode"
             name="zipCode"
             label="Zip-code"
@@ -121,7 +121,7 @@
           class="form__element"
         >
           <SfSelect
-            data-cy="shipping-details-select_country"
+            data-cy="billing-details-select_country"
             class="form__select sf-select--underlined"
             :value="form.country"
             @selected="form.country = $event"
@@ -147,7 +147,7 @@
         class="form__element"
       >
         <SfInput
-          data-cy="shipping-details-input_phoneNumber"
+          data-cy="billing-details-input_phoneNumber"
           v-model="form.phone"
           name="phone"
           label="Phone number"
@@ -156,7 +156,7 @@
           :errorMessage="errors[0]"
         />
       </ValidationProvider>
-      <SfButton data-cy="shipping-details-btn_update" class="form__button">
+      <SfButton data-cy="billing-details-btn_update" class="form__button">
         {{ isNew ? "Add the address" : "Update the address" }}
       </SfButton>
     </form>
@@ -191,7 +191,7 @@ extend('oneOf', {
 });
 
 export default {
-  name: 'ShippingAddressForm',
+  name: 'BillingAddressForm',
 
   components: {
     SfInput,