+++ packages/core/cli/__tests__/index.spec.ts
@@ -1,4 +1,4 @@
-import { cli } from '../src/index';
+import { cli } from '@vue-storefront/cli/src/index';
 import log from '@vue-storefront/cli/src/utils/log';
 import init from '../src/commands/init';