+++ packages/core/cli/__tests__/utils/getIntegrations.spec.ts
@@ -1,4 +1,4 @@
-import getIntegrations from '../../src/utils/getIntegrations';
+import getIntegrations from '@vue-storefront/cli/src/utils/getIntegrations';
 
 const expectedIntegrations = ['commercetools', 'shopify'];