+++ packages/core/cli/src/scripts/createProject/copyIntegrationTheme.ts
@@ -1,9 +1,9 @@
 import path from 'path';
-import { copyThemeFiles, getThemePath } from '../../utils/helpers';
+import { copyThemeFiles, getDependencyPath } from '../../utils/helpers';
 import fs from 'fs';
 
 export default async (integration: string, targetPath: string, omitFiles: Array<string> = []): Promise<void> => {
-  const integrationThemePath = getThemePath(`${integration}-theme`);
+  const integrationThemePath = getDependencyPath(`${integration}-theme`);
   const integrationThemeFiles = fs.readdirSync(integrationThemePath)
     .filter(fileName => !omitFiles.includes(fileName))
     .map(directory => path.join(integrationThemePath, directory));