+++ packages/core/core/src/factories/useWishlistFactory.ts
@@ -1,18 +1,18 @@
-import { UseWishlist } from '../types';
+import { UseWishlist, CustomQuery } from '../types';
 import { Ref, computed } from '@vue/composition-api';
 import { sharedRef, Logger } from '../utils';
 
 export type UseWishlistFactoryParams<WISHLIST, WISHLIST_ITEM, PRODUCT> = {
-  loadWishlist: () => Promise<WISHLIST>;
+  loadWishlist: (customQuery?: CustomQuery) => Promise<WISHLIST>;
   addToWishlist: (params: {
     currentWishlist: WISHLIST;
     product: PRODUCT;
-  }) => Promise<WISHLIST>;
+  }, customQuery?: CustomQuery) => Promise<WISHLIST>;
   removeFromWishlist: (params: {
     currentWishlist: WISHLIST;
     product: WISHLIST_ITEM;
-  }) => Promise<WISHLIST>;
-  clearWishlist: (prams: { currentWishlist: WISHLIST }) => Promise<WISHLIST>;
+  }, customQuery?: CustomQuery) => Promise<WISHLIST>;
+  clearWishlist: (params: { currentWishlist: WISHLIST }) => Promise<WISHLIST>;
   isOnWishlist: (params: { currentWishlist: WISHLIST; product: PRODUCT }) => boolean;
 };
 
@@ -24,46 +24,52 @@ interface UseWishlistFactory<WISHLIST, WISHLIST_ITEM, PRODUCT> {
 export const useWishlistFactory = <WISHLIST, WISHLIST_ITEM, PRODUCT>(
   factoryParams: UseWishlistFactoryParams<WISHLIST, WISHLIST_ITEM, PRODUCT>
 ): UseWishlistFactory<WISHLIST, WISHLIST_ITEM, PRODUCT> => {
-  const loading: Ref<boolean> = sharedRef<boolean>(false, 'useWishlist-loading');
-  const wishlist: Ref<WISHLIST> = sharedRef(null, 'useWishlist-wishlist');
-
   const setWishlist = (newWishlist: WISHLIST) => {
-    wishlist.value = newWishlist;
+    sharedRef('useWishlist-wishlist').value = newWishlist;
+    Logger.debug('useWishlistFactory.setWishlist', newWishlist);
   };
 
   const useWishlist = (): UseWishlist<WISHLIST, WISHLIST_ITEM, PRODUCT> => {
+    const loading: Ref<boolean> = sharedRef<boolean>(false, 'useWishlist-loading');
+    const wishlist: Ref<WISHLIST> = sharedRef(null, 'useWishlist-wishlist');
 
-    const addToWishlist = async (product: PRODUCT) => {
+    const addToWishlist = async (product: PRODUCT, customQuery?: CustomQuery) => {
       Logger.debug('useWishlist.addToWishlist', product);
 
       loading.value = true;
-      const updatedWishlist = await factoryParams.addToWishlist({
-        currentWishlist: wishlist.value,
-        product
-      });
+      const updatedWishlist = await factoryParams.addToWishlist(
+        {
+          currentWishlist: wishlist.value,
+          product
+        },
+        customQuery
+      );
       wishlist.value = updatedWishlist;
       loading.value = false;
     };
 
-    const removeFromWishlist = async (product: WISHLIST_ITEM) => {
+    const removeFromWishlist = async (product: WISHLIST_ITEM, customQuery?: CustomQuery) => {
       Logger.debug('useWishlist.removeFromWishlist', product);
 
       loading.value = true;
-      const updatedWishlist = await factoryParams.removeFromWishlist({
-        currentWishlist: wishlist.value,
-        product
-      });
+      const updatedWishlist = await factoryParams.removeFromWishlist(
+        {
+          currentWishlist: wishlist.value,
+          product
+        },
+        customQuery
+      );
       wishlist.value = updatedWishlist;
       loading.value = false;
     };
 
-    const loadWishlist = async () => {
+    const loadWishlist = async (customQuery?: CustomQuery) => {
       Logger.debug('useWishlist.loadWishlist');
 
       if (wishlist.value) return;
 
       loading.value = true;
-      wishlist.value = await factoryParams.loadWishlist();
+      wishlist.value = await factoryParams.loadWishlist(customQuery);
       loading.value = false;
     };