+++ packages/core/nuxt-theme-module/theme/components/UserShippingAddress.vue
@@ -1,9 +1,9 @@
 <template>
-  <b>Please implement vendor specific UserAddress component in 'components' directory</b>
+  <b>Please implement vendor specific UserShippingAddress component in 'components' directory</b>
 </template>
 
 <script>
 export default {
-  name: 'UserAddress'
+  name: 'UserShippingAddress'
 };
 </script>